+++ src/main/java/com/domainevent/sample/db/JsonDomainEventType.java
@@ -1,0 +1,4 @@
+package com.domainevent.sample.db;
+
+public class JsonDomainEventType {
+}
