+++ src/main/java/com/domainevent/sample/domain/events/DomainEvent.java
@@ -2,6 +2,7 @@ package com.domainevent.sample.domain.events;
 
 import java.time.Instant;
 
+
 public interface DomainEvent {
 
     Instant occurredOn();
